重构预览模块
