+++ paas-ce/lesscode/lib/client/src/preview/children/home.vue
@@ -1,0 +1,8 @@
+<template>
+    <router-view></router-view>
+</template>
+<script>
+    export default {
+        name: 'Home'
+    }
+</script>
